Revert "remove P.DateRecurringIndex again after p.a.event merge, since it's automatically loaded by Zope"
Apparently it is not automatically loaded by Zope, or something else
is going wrong, as this caused testing failures, especially the
strange 'AttributeError: adapters', signalling that the component
registry was not setup correctly.  Ultimately this was a ValueError
during creation of a Plone Site:

ValueError: Index of type DateRecurringIndex not found

For clarity: the removal was reverted, so Products.DateRecurringIndex
is now loaded again.

This reverts commit 13d1271ffb914d9001dbfe91bdb01bb7604f972c.