No longer try to load `Products.SecureMailHost` and its zcml.
This is not shipped with Plone 5.0 or higher.
Depending on the import order in tests, there were two possibilities:

1. Products.SecureMailHost could not be imported.
2. Products.SecureMailHost was a fake module from Products.CMFPlone.patches.securemailhost.

In the first case, there was nothing for us to do.
In the second case, test setup completely failed, with:
AttributeError: 'module' object has no attribute 'SecureMailHost'

So: it is useless to still try to load this.