[fix] Preserve ZCA site for transaction hooks that depend on it
The `zope.component` site should not be cleared before committing transacions
in case any transaction hooks depend on it.

In the test in which I ran into this issue, the `transaction.commit()` failed
because the indexing queue was processing a
`plone.app.blocks.indexing.LayoutSearchableText` indexer at commit-time that
was trying to lookup a `plone.app.blocks.layoutbehavior.ILayoutAware` adapter.
Since that adapter depends on behavior-derived interfaces,
`plone.dexterity.schema` looks up a `plone.dexterity.interfaces.IDexterityFTI`
utility for the `portal_type` which fails if the site isn't set which in turn
yields no behavior interfaces which in turn makes the behavior-based adapter
lookup fail.

In my case, this wasn't even the symptom observed.  Because the commit failed
silently, the changes from applying the GS profile weren't persisted
including, in my case, a custom theme browser layer registration which in turn
meant a custom browser view lookup failed.  IOW, this silent commit failure
can lead to any number of very surprising, very wrong, and hard to diagnose
behaviors.

I'm not sure what changed or when, but this test worked under Python 2.7 and
Plone 5.1 and I'm fairly confident that nothing changed in the testing code on
"my" side that caused this and that the root change in behavior is somehow on
the Plone side of things.